Report on test failures not caused by asserts
Not all tests fail as a result of Xunit assertions. An example of this
is
`AboutNull.AWayNotToCheckThatAnObjectIsNull()` which throws a
`NullReferenceException` before the assertion can be checked. The test
failure is squelched and silently ignored as a result of the
unsuccessful soft cast to the expected `XunitException`. By removing
the cast we can catch any test failure while preserving relevant
reporting details.