Adds AboutGlobalization to Enlightment Path (#69)
Adds the AboutGlobalization class to the path to enlightenment. This
class consists of 4 learning steps which explore how strings are
handled in an international sense. It looks at how each character in
a string can consist of 1 or more Char objects, explores how to test
strings for equality (when to use Ordinal vs Culture specific) and
how strings are sorted based on culture.